partial reversion: use same error message for reset and apply buttons
This allows one to select a filter that returns zero words if the
filter applies to some words in the dictionary but all those words have
already been completed and removed from the word list. This should be
okay though; it will just mean that the user sees a pre-emptive "all
done!" message. Need to pass "completed" word array to the Settings
component if I want to avoid this.